Extract abstract super class for TestHiveRoles
